# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentação Microserviço agenda médica",
        "description": "Swagger Microserviço agenda médica",
        "contact": {
            "email": "1094820@sga.pucminas.br"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/consulta": {
            "get": {
                "tags": [
                    "Consulta"
                ],
                "summary": "Retorna todas as consultas",
                "description": "Retorna todas as consultas",
                "operationId": "9ecd90faa00fa687d1a688f003421b93",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Consulta"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Não autenticado",
                                        "value": {
                                            "message": "unauthenticated"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Mensagem de erro da exception",
                                        "value": {
                                            "message": "Exemplo de mensagem de erro"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Consulta"
                ],
                "summary": "Cadastra nova consulta",
                "description": "Cadastra nova consulta",
                "operationId": "bdb53333c035b842854c6badeba79304",
                "requestBody": {
                    "description": "Dados da consulta",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Consulta"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cadastrado com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Consulta"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Não autenticado",
                                        "value": {
                                            "message": "unauthenticated"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Entrada de dados inválida.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Dados inválidos",
                                        "value": {
                                            "message": "Dados inválidos."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Mensagem de erro da exception",
                                        "value": {
                                            "message": "Exemplo de mensagem de erro"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        },
        "/consulta/{id}": {
            "get": {
                "tags": [
                    "Consulta"
                ],
                "summary": "Retorna consulta pelo id",
                "description": "Retorna consulta pelo id",
                "operationId": "d33a8831280efdc53a736a965d2ba87b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da consulta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Consulta"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Não autenticado",
                                        "value": {
                                            "message": "unauthenticated"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Entrada de dados inválida.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Dados inválidos",
                                        "value": {
                                            "message": "Dados inválidos."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Mensagem de erro da exception",
                                        "value": {
                                            "message": "Exemplo de mensagem de erro"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Consulta"
                ],
                "summary": "Edita consulta",
                "description": "Edita consulta",
                "operationId": "5b9a8ecd866bd1e14d36e4dddda706f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da consulta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Dados da consulta a serem atualizados",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Consulta"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Salvo com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Consulta"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Não autenticado",
                                        "value": {
                                            "message": "unauthenticated"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Entrada de dados inválida.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Dados inválidos",
                                        "value": {
                                            "message": "Dados inválidos."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Mensagem de erro da exception",
                                        "value": {
                                            "message": "Exemplo de mensagem de erro"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Consulta"
                ],
                "summary": "Apaga registro da consulta",
                "description": "Apaga registro da consulta",
                "operationId": "7b2e07a54a49bd5b77f99422e46bdcc9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id da consulta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Apagado com sucesso."
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Não autenticado",
                                        "value": {
                                            "message": "unauthenticated"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Entrada de dados inválida.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Dados inválidos",
                                        "value": {
                                            "message": "Dados inválidos."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Mensagem de erro da exception",
                                        "value": {
                                            "message": "Exemplo de mensagem de erro"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        },
        "/consulta/paciente/{id}": {
            "get": {
                "tags": [
                    "Consulta"
                ],
                "summary": "Recupera as consultas de um paciente",
                "description": "Recupera as consultas de um paciente",
                "operationId": "6fba6ab6678aada8fbbc9a8c81dea577",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do paciente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Consultas.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Consulta"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autenticado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Não autenticado",
                                        "value": {
                                            "message": "unauthenticated"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno do servidor.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Mensagem de erro da exception",
                                        "value": {
                                            "message": "Exemplo de mensagem de erro"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Consulta": {
                "title": "Consulta",
                "properties": {
                    "data_hora": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "tipo": {
                        "type": "string"
                    },
                    "observacoes": {
                        "type": "string"
                    },
                    "fk_medico_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "fk_paciente_id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "fk_unidade_atendimento_id": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api_token": {
                "type": "apiKey",
                "description": "Entre com o token beraer no formato **Bearer &lt;token>**",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}